+++ CHANGELOG.md
@@ -5,8 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Flattened module structure: removed `protocol` module, types now at crate root.
+
 ## [0.1.0] - 2026-03-06
 
 Initial release.
 
+[Unreleased]: https://github.com/mbr/mercutio/compare/0.1.0...HEAD
 [0.1.0]: https://github.com/mbr/mercutio/releases/tag/0.1.0